discord_bot/pyproject.toml
[project]
name = "ladlorchart-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
]
